+++ service/controller/managementcluster/resource.go
@@ -127,7 +127,6 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 			K8sClient:      config.K8sClient,
 			Logger:         config.Logger,
 			Installation:   config.Installation,
-			Provider:       config.Provider,
 			Proxy:          config.Proxy,
 			OpsgenieKey:    config.OpsgenieKey,
 			GrafanaAddress: config.GrafanaAddress,