+++ service/controller/resource/alerting/alertmanagerconfig/resource.go
@@ -12,7 +12,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/cluster"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/template"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
@@ -30,7 +29,6 @@ type Config struct {
 	Logger    micrologger.Logger
 
 	Installation   string
-	Provider       cluster.Provider
 	Proxy          func(reqURL *url.URL) (*url.URL, error)
 	OpsgenieKey    string
 	GrafanaAddress string
@@ -43,7 +41,6 @@ type NotificationTemplateData struct {
 }
 
 type AlertmanagerTemplateData struct {
-	Provider       string
 	Installation   string
 	ProxyURL       string
 	OpsgenieKey    string
@@ -150,7 +147,6 @@ func getTemplateData(config Config) (*AlertmanagerTemplateData, error) {
 	}
 
 	d := &AlertmanagerTemplateData{
-		Provider:       config.Provider.Kind,
 		Installation:   config.Installation,
 		OpsgenieKey:    config.OpsgenieKey,
 		GrafanaAddress: config.GrafanaAddress,